requirements.txt
httpx>=0.27,<1.0
python-dotenv>=1.0,<2.0
websockets>=12.0,<13.0
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
jinja2>=3.1,<4.0
tzdata>=2024.1